requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
dj-database-url==2.1.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
# Removed boto3 and django-storages (using local storage with fly.io volumes)
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0
django-filter==23.5
djangorestframework-simplejwt==5.3.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.26
requests==2.31.0

# Additional components from plan
drf-spectacular==0.26.5
graphene-django==3.2.0
django-graphql-jwt==0.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-prometheus==2.3.1
django-health-check==3.17.0
django-simple-history==3.4.0

# GeoDjango dependencies  
GDAL==3.4.3
Fiona==1.9.0

# Tracking system dependencies
djangorestframework-gis==1.0
